move Either to bokeh.core.property.either
